refactor: 날짜 계산 로직 수정
